jules: feat: add LLM Audience Copilot to presenter UI
- Wraps base LLM adapter with `CopilotLLM` struct to intercept completions.
- Checks if the transcript is an audience question via `IsQuestion`.
- Bypasses LLM generation entirely if not a question to save compute.
- Appends loaded script content as context for accurate Q&A responses.
- Updates Svelte frontend with a glowing bottom-right Copilot drawer tied to Wails `thinking` events.
- Reaches 94%+ test coverage for the updated presenter package.

+++ wails-app/products/presenter/app_test.go
@@ -14,7 +14,7 @@ import (
 
 // Define dummy cache adapter so we don't import the one that needs CGO
 type DummyCache struct{}
-func (d *DummyCache) Search(embedding []float32, threshold float32) (string, bool) { return "", false }
+func (d *DummyCache) Search(embedding []float32, threshold float64) (string, bool) { return "", false }
 func (d *DummyCache) Store(question, answer string) error { return nil }
 func (d *DummyCache) Count() int { return 0 }
 func (d *DummyCache) Close() error { return nil }
